// src/models/beers.ts
// Generated by https://quicktype.io

export interface IBeer {
	product_id: number;
	name: string;
	size: string;
	price: string;
	beer_id: number;
	image_url: string;
	category: string;
	abv: string;
	style: string;
	attributes: string;
	type: string;
	brewer: string;
	country: string;
	on_sale: boolean;
}